moved classes located on previous root folder to new.
