+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/v3/VendorExtendable.java
